Added last algorithm
Segmentation fault!

+++ algorithm/Tile.h
@@ -6,16 +6,16 @@
 #include "TileDirection.h"
 
 // Bits 0-3 are reserved for the walls. 
-constexpr uint8_t kVictimBit = 4;
-constexpr uint8_t kObstacleBit = 5;
-constexpr uint8_t kBlackTileBit = 6;
-constexpr uint8_t kCheckpointBit = 7;
+constexpr int kVictimBit = 4;
+constexpr int kObstacleBit = 5;
+constexpr int kBlackTileBit = 6;
+constexpr int kCheckpointBit = 7;
 
-constexpr uint8_t kNumberOfDirections = 4;
+constexpr int kNumberOfDirections = 4;
 
-constexpr uint8_t kWhiteTileWeight = 1;// min weight
-constexpr uint8_t kBlueTileWeight = 3; // 5 seconds
-constexpr uint8_t kRampWeight = 5;
+constexpr int kWhiteTileWeight = 1;// min weight
+constexpr int kBlueTileWeight = 3; // 5 seconds
+constexpr int kRampWeight = 5;
 
 constexpr coord kInvalidPosition = coord{1000,1000,1000};
